pyproject.toml
# SPDX-FileCopyrightText: 2025 Rubber Duck Contributors
# SPDX-License-Identifier: MPL-2.0

[project]
name = "rubber-duck"
version = "0.1.0"
description = "A Discord bot that nudges you with context-aware prompts and remembers everything"
requires-python = ">=3.11"
license = { text = "MPL-2.0" }
authors = [
    { name = "Luis Villa", email = "luis@lu.is" }
]

dependencies = [
    "discord.py>=2.3.0",
    "anthropic>=0.40.0",
    "apscheduler>=3.10.0",
    "letta-client>=1.6.0",
    "pyyaml>=6.0",
    "todoist-api-python>=3.1.0",
]

[project.optional-dependencies]
gcal = [
    "google-auth>=2.0.0",
    "google-api-python-client>=2.0.0",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.1.0",
    "google-auth>=2.0.0",
    "google-api-python-client>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.hatch.build.targets.wheel]
packages = ["src/rubber_duck"]
